Remove NominalVoltageStyleProvider inheritance of TopologicalStyleProvider
